+++ src/components/landing/AboutMe/index.tsx
@@ -1,0 +1,16 @@
+import './style.scss'
+import Skill from './Skill'
+
+function Skills() {
+  return (
+    <div>
+    <Skill
+      name='Front-end'
+      detail='React, Typescript'
+      icon={reactIcon}
+    />
+    </div>
+  )
+}
+
+export default Skills
