--- src/components/landing/Skills/index.tsx
@@ -1,16 +1,0 @@
-import './style.scss'
-import Skill from './Skill'
-
-function Skills() {
-  return (
-    <div>
-    <Skill
-      name='Front-end'
-      detail='React, Typescript'
-      icon={reactIcon}
-    />
-    </div>
-  )
-}
-
-export default Skills
