+++ src/styles/_main.scss
@@ -1,0 +1,2 @@
+@import "colors";
+@import "fonts"
